pyproject.toml
[project]
name = "brewfather_mcp"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
